+++ asm/nonmatchings/code_23680/func_800494C0.s
@@ -1,0 +1,42 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_800494C0
+/* 248C0 800494C0 27BDFFD8 */  addiu     $sp, $sp, -0x28
+/* 248C4 800494C4 AFB1001C */  sw        $s1, 0x1c($sp)
+/* 248C8 800494C8 0080882D */  daddu     $s1, $a0, $zero
+/* 248CC 800494CC AFB20020 */  sw        $s2, 0x20($sp)
+/* 248D0 800494D0 00A0902D */  daddu     $s2, $a1, $zero
+/* 248D4 800494D4 AFBF0024 */  sw        $ra, 0x24($sp)
+/* 248D8 800494D8 AFB00018 */  sw        $s0, 0x18($sp)
+/* 248DC 800494DC 822400A4 */  lb        $a0, 0xa4($s1)
+/* 248E0 800494E0 0C00FB3A */  jal       get_enemy
+/* 248E4 800494E4 00C0802D */   daddu    $s0, $a2, $zero
+/* 248E8 800494E8 3C03FFFF */  lui       $v1, 0xffff
+/* 248EC 800494EC 02031824 */  and       $v1, $s0, $v1
+/* 248F0 800494F0 34650002 */  ori       $a1, $v1, 2
+/* 248F4 800494F4 32100001 */  andi      $s0, $s0, 1
+/* 248F8 800494F8 12000003 */  beqz      $s0, .L80049508
+/* 248FC 800494FC 0040182D */   daddu    $v1, $v0, $zero
+/* 24900 80049500 3C020001 */  lui       $v0, 1
+/* 24904 80049504 00A22825 */  or        $a1, $a1, $v0
+.L80049508:
+/* 24908 80049508 8C620018 */  lw        $v0, 0x18($v1)
+/* 2490C 8004950C 9442002A */  lhu       $v0, 0x2a($v0)
+/* 24910 80049510 30420020 */  andi      $v0, $v0, 0x20
+/* 24914 80049514 10400002 */  beqz      $v0, .L80049520
+/* 24918 80049518 3C020002 */   lui      $v0, 2
+/* 2491C 8004951C 00A22825 */  or        $a1, $a1, $v0
+.L80049520:
+/* 24920 80049520 C6200040 */  lwc1      $f0, 0x40($s1)
+/* 24924 80049524 E7A00010 */  swc1      $f0, 0x10($sp)
+/* 24928 80049528 8E260038 */  lw        $a2, 0x38($s1)
+/* 2492C 8004952C 8E27003C */  lw        $a3, 0x3c($s1)
+/* 24930 80049530 0C052757 */  jal       play_sound_at_position
+/* 24934 80049534 0240202D */   daddu    $a0, $s2, $zero
+/* 24938 80049538 8FBF0024 */  lw        $ra, 0x24($sp)
+/* 2493C 8004953C 8FB20020 */  lw        $s2, 0x20($sp)
+/* 24940 80049540 8FB1001C */  lw        $s1, 0x1c($sp)
+/* 24944 80049544 8FB00018 */  lw        $s0, 0x18($sp)
+/* 24948 80049548 03E00008 */  jr        $ra
+/* 2494C 8004954C 27BD0028 */   addiu    $sp, $sp, 0x28
