+++ asm/nonmatchings/code_23680/func_80049ECC.s
@@ -1,0 +1,51 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_80049ECC
+/* 252CC 80049ECC 27BDFFE8 */  addiu     $sp, $sp, -0x18
+/* 252D0 80049ED0 AFB00010 */  sw        $s0, 0x10($sp)
+/* 252D4 80049ED4 0080802D */  daddu     $s0, $a0, $zero
+/* 252D8 80049ED8 AFBF0014 */  sw        $ra, 0x14($sp)
+/* 252DC 80049EDC 8E020148 */  lw        $v0, 0x148($s0)
+/* 252E0 80049EE0 0C00EABB */  jal       get_npc_unsafe
+/* 252E4 80049EE4 84440008 */   lh       $a0, 8($v0)
+/* 252E8 80049EE8 0040202D */  daddu     $a0, $v0, $zero
+/* 252EC 80049EEC C480001C */  lwc1      $f0, 0x1c($a0)
+/* 252F0 80049EF0 44801000 */  mtc1      $zero, $f2
+/* 252F4 80049EF4 44801800 */  mtc1      $zero, $f3
+/* 252F8 80049EF8 46000021 */  cvt.d.s   $f0, $f0
+/* 252FC 80049EFC 4622003E */  c.le.d    $f0, $f2
+/* 25300 80049F00 00000000 */  nop
+/* 25304 80049F04 45000009 */  bc1f      .L80049F2C
+/* 25308 80049F08 0000182D */   daddu    $v1, $zero, $zero
+/* 2530C 80049F0C C480003C */  lwc1      $f0, 0x3c($a0)
+/* 25310 80049F10 C4820064 */  lwc1      $f2, 0x64($a0)
+/* 25314 80049F14 4602003E */  c.le.s    $f0, $f2
+/* 25318 80049F18 00000000 */  nop
+/* 2531C 80049F1C 45000003 */  bc1f      .L80049F2C
+/* 25320 80049F20 00000000 */   nop
+/* 25324 80049F24 E482003C */  swc1      $f2, 0x3c($a0)
+/* 25328 80049F28 24030001 */  addiu     $v1, $zero, 1
+.L80049F2C:
+/* 2532C 80049F2C 14600009 */  bnez      $v1, .L80049F54
+/* 25330 80049F30 2403F7FF */   addiu    $v1, $zero, -0x801
+/* 25334 80049F34 C484003C */  lwc1      $f4, 0x3c($a0)
+/* 25338 80049F38 C480001C */  lwc1      $f0, 0x1c($a0)
+/* 2533C 80049F3C 46002100 */  add.s     $f4, $f4, $f0
+/* 25340 80049F40 C4820014 */  lwc1      $f2, 0x14($a0)
+/* 25344 80049F44 46020001 */  sub.s     $f0, $f0, $f2
+/* 25348 80049F48 E484003C */  swc1      $f4, 0x3c($a0)
+/* 2534C 80049F4C 080127DB */  j         .L80049F6C
+/* 25350 80049F50 E480001C */   swc1     $f0, 0x1c($a0)
+.L80049F54:
+/* 25354 80049F54 8C820000 */  lw        $v0, ($a0)
+/* 25358 80049F58 AC80001C */  sw        $zero, 0x1c($a0)
+/* 2535C 80049F5C 00431024 */  and       $v0, $v0, $v1
+/* 25360 80049F60 AC820000 */  sw        $v0, ($a0)
+/* 25364 80049F64 2402000C */  addiu     $v0, $zero, 0xc
+/* 25368 80049F68 AE020070 */  sw        $v0, 0x70($s0)
+.L80049F6C:
+/* 2536C 80049F6C 8FBF0014 */  lw        $ra, 0x14($sp)
+/* 25370 80049F70 8FB00010 */  lw        $s0, 0x10($sp)
+/* 25374 80049F74 03E00008 */  jr        $ra
+/* 25378 80049F78 27BD0018 */   addiu    $sp, $sp, 0x18
