+++ asm/nonmatchings/code_23680/get_coin_drop_amount.s
