+++ asm/nonmatchings/code_2BF90/func_80050E18.s
@@ -1,0 +1,33 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_80050E18
+/* 2C218 80050E18 00041100 */  sll       $v0, $a0, 4
+/* 2C21C 80050E1C 00441023 */  subu      $v0, $v0, $a0
+/* 2C220 80050E20 000210C0 */  sll       $v0, $v0, 3
+/* 2C224 80050E24 00441021 */  addu      $v0, $v0, $a0
+/* 2C228 80050E28 00021080 */  sll       $v0, $v0, 2
+/* 2C22C 80050E2C 3C03800A */  lui       $v1, %hi(D_8009A628)
+/* 2C230 80050E30 8C63A628 */  lw        $v1, %lo(D_8009A628)($v1)
+/* 2C234 80050E34 24420024 */  addiu     $v0, $v0, 0x24
+/* 2C238 80050E38 00621821 */  addu      $v1, $v1, $v0
+/* 2C23C 80050E3C 8C620004 */  lw        $v0, 4($v1)
+/* 2C240 80050E40 1040000E */  beqz      $v0, .L80050E7C
+/* 2C244 80050E44 00000000 */   nop
+/* 2C248 80050E48 8C620008 */  lw        $v0, 8($v1)
+/* 2C24C 80050E4C 1040000B */  beqz      $v0, .L80050E7C
+/* 2C250 80050E50 00000000 */   nop
+/* 2C254 80050E54 10A00008 */  beqz      $a1, .L80050E78
+/* 2C258 80050E58 24020003 */   addiu    $v0, $zero, 3
+/* 2C25C 80050E5C 240200FF */  addiu     $v0, $zero, 0xff
+/* 2C260 80050E60 A062002A */  sb        $v0, 0x2a($v1)
+/* 2C264 80050E64 24020002 */  addiu     $v0, $zero, 2
+/* 2C268 80050E68 A4650028 */  sh        $a1, 0x28($v1)
+/* 2C26C 80050E6C A060002B */  sb        $zero, 0x2b($v1)
+/* 2C270 80050E70 03E00008 */  jr        $ra
+/* 2C274 80050E74 A0620026 */   sb       $v0, 0x26($v1)
+.L80050E78:
+/* 2C278 80050E78 A0620024 */  sb        $v0, 0x24($v1)
+.L80050E7C:
+/* 2C27C 80050E7C 03E00008 */  jr        $ra
+/* 2C280 80050E80 00000000 */   nop
