+++ asm/nonmatchings/os/code_4a1f0_len_170/func_8006EF20.s
@@ -1,0 +1,6 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_8006EF20
+/* 4A320 8006EF20 03E00008 */  jr        $ra
+/* 4A324 8006EF24 00000000 */   nop
