+++ asm/nonmatchings/os/code_4a1f0_len_170/func_8006EF58.s
@@ -1,0 +1,6 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_8006EF58
+/* 4A358 8006EF58 03E00008 */  jr        $ra
+/* 4A35C 8006EF5C 00000000 */   nop
