+++ asm/nonmatchings/os/code_4a360_len_930/func_8006EF60.s
@@ -1,0 +1,6 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_8006EF60
+/* 4A360 8006EF60 03E00008 */  jr        $ra
+/* 4A364 8006EF64 00000000 */   nop
