+++ asm/nonmatchings/os/code_4ac90_len_3910/func_80072CEC.s
@@ -1,0 +1,6 @@
+.set noat      # allow manual use of $at
+.set noreorder # don't insert nops after branches
+
+glabel func_80072CEC
+/* 4E0EC 80072CEC 03E00008 */  jr        $ra
+/* 4E0F0 80072CF0 00000000 */   nop
