+++ src/code_25AF0.c
@@ -1,0 +1,7 @@
+#include "common.h"
+
+INCLUDE_ASM(s32, "code_25AF0", func_8004A6F0);
+
+INCLUDE_ASM(s32, "code_25AF0", func_8004A73C);
+
+INCLUDE_ASM(s32, "code_25AF0", func_8004A784);
