+++ tools/splat.yaml
@@ -62,9 +62,12 @@ segments:
     - [0x1a1f0, c, code_1a1f0_len_5390]
     - [0x1f580, c, code_1f580_len_1940]
     - [0x20ec0, c, code_20ec0_len_5040]
+    - [0x23680, c]
+    - [0x25AF0, c]
     - [0x25f00, c, code_25f00_len_940]
     - [0x26840, c, code_26840_len_20d0]
     - [0x28910, c, code_28910_len_5090]
+    - [0x2BF90, c]
     - [0x2d9a0, c, code_2d9a0_len_890]
     - [0x2e230, c, code_2e230_len_2190]
     - [0x303c0, c, code_303c0_len_3e10]
